+++ petsadv/petsquare-backend/petsquarebackend/models/image.py
@@ -31,13 +31,13 @@ import base64
 
 class Image_TB(Base):
     __tablename__ = 'image'
-    id              = Column(Integer(10), nullable=False, unique=True, 
+    id              = Column(Integer, nullable=False, unique=True, 
                              primary_key=True, autoincrement=True)
     description     = Column(String(255), nullable=True, unique=False,)
     filename        = Column(String(255), nullable=False, unique=False,)
     image           = Column(BLOB)
     format          = Column(String(20), nullable=False, unique=False,)
-    userid          = Column(Integer(10), nullable=False, unique=False,)
+    userid          = Column(Integer, nullable=False, unique=False,)
     createddatetime = Column(DateTime, nullable=False)
     updateddatetime = Column(DateTime, nullable=False)
     checks    = relationship('Check_TB',  backref=backref('image', order_by=id))
